added a start screen with a simple menu
in program.cs file I added the StartScreenMenu instead of the FrmLevel.cs so that on starting the game the start screen shows up instead of the game. Then I linked the FrmLevel.cs to the start button in the StartScreenMenu.cs file, so that on clicking start button the game shows up. Added a background for the start screen. Added Help and Settings button but they don't do anything.